Update power-of-thor-episode-1.js
analysis of failure and rethinking a solution. still stuck on test 3. but ThorX ThorY does not corelate with the output location of thor at the time of death. 
I think I need to redo the code with the localized coordinates of 0,0 at the top left instead of go North = +Y